Reply id for sphinx ProcessingResult
This is a hack and should be fixed at some point (FIXME):
the reply_id consists of the last blinded dh_pubkey of the source,
which is also returned by the compute_shared_keys function in the
sphinx_end_host module. This allows the user to recognize exacly to
what header it previously constructed an incoming reply belongs.