Blinding factors and reply id for session requests
Added the blinding factors, since because of limitations in the
curve25519 library its not possible to compute a new exponent
- a Private - from two exponents, so it is not possible to compute
all the source's temporary private keys resulting of the successive
blindings. Instead, I resorted to storing only the first private key
and all the blinding factors.
The reply id will be the blinded dh_pubkey as seen by the last hop
(the source), since this value is returned by the function
compute_shared_keys in the sphinx_end_host module, and is also used
to recognize a reply by adding it to the list of expected replies
(see function SphinxEndHost.add_expected_reply).