Inverse of nonce derivation, changed size to 16
Changed function derive_new_nonce so that it is invertible, created
the inverse function derive_previous_nonce. To do this, the size of
the nonce was changed to Crypto.Cipher.AES.block_size, which is 16.
If a shorter nonce is to be used, some FPE scheme could be used, but
it seems there is no widely used solution for this (what is needed is
a block cipher with block size as small as the nonce).