Incoming mac and blinded header now in SessionInfo
To support fast incoming packet verification on the source, the
incoming mac and blinded header, previously stored only in the
DestinationSessionInfo class, were moved to SessionInfo, so that they
are available also to SourceSessionInfo.
Changed the _create_basic_blinded_header function in HornetSource so
that it also returns the last mac and blinded header, so that for the
backward path the source can store them in the SessionInfo instance.