Add a direction flag to sibra reservation info blocks
When a leaf AS registers a sibra steady path with a core AS, it needs to
set a flag to say that the opaque-fields should be verified in reverse
order.

Also update sibra_ext_test.py to look up steady paths between two
specified ASes, and test that it works.